[Upd #115] Refactor for readability
