+++ src/stories/Answer.stories.tsx
@@ -1,17 +1,18 @@
 import React from "react";
 import Answer from "../components/Answer";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import question from "./assets/questionWithMedia.json";
-import questionTypeHead from "./assets/questionTypeHead.json";
-import questionDate from "./assets/questionDate.json";
-import questionCheckBox from "./assets/questionCheckBox.json";
-import questionMaskedInput from "./assets/questionMaskedInput.json";
-import questionSparqlInput from "./assets/questionSparqlInput.json";
-import questionTurtleInput from "./assets/questionTurtleInput.json";
 import { ConfigurationContextProvider } from "../contexts/ConfigurationContext";
 import Constants from "../constants/Constants";
 import { FormGenContextProvider } from "../contexts/FormGenContext";
+
 import possibleValues from "./assets/possibleValues.json";
+import question from "./assets/question/questionWithMedia.json";
+import questionTypeHead from "./assets/question/questionTypeHead.json";
+import questionDate from "./assets/question/questionDate.json";
+import questionCheckBox from "./assets/question/questionCheckBox.json";
+import questionMaskedInput from "./assets/question/questionMaskedInput.json";
+import questionSparqlInput from "./assets/question/questionSparqlInput.json";
+import questionTurtleInput from "./assets/question/questionTurtleInput.json";
 
 export default {
   title: "Components/Answer",
