+++ src/stories/assets/question/questionCheckBox.json
