+++ src/stories/assets/question/questionMaskedInput.json
