+++ src/stories/assets/question/questionWithMedia.json
